+++ CHANGES.md
@@ -1,6 +1,10 @@
-## In the latest release...
+## In the next release...
 
 * Expire idle services and clients.
+* Convert `thriftProtocol` from a client/server param to a router param.
+* Add support for client auth TLS.
+* Add TLS support for `io.l5d.httpController` and `io.l5d.mesh`.
+* Add support for HTTP/2 tracing.
 * Cache dtab observations in the io.l5d.consul store.
 
 ## 1.0.2 2017-05-12